fix: add initial scan guessing component
Added a mutex lock to the camera analyzer response so that there is not
more than one image being analyzed at a time.

Added 'MatchCameraImageUseCase' to embed the image, call the supabase
query and release the lock

Made the analyzer pre-rotate if necessary so that the ui is always
working with the same image.

TODO: Incomplete commit

Needs UI and animations for matching an image.
Needs UI for no match, keep trying
Needs to bring in the other clues to help guide the user (color, label,
location)