fix: add suspend invoke capability to stores
As an alternative to dispatchers, stores can now invoke and suspend
effects for a result

Used by the camera scanner to suspend for a another image.

TODO:  the detections and segmentations need user interaction to specify
sub-images to run analyzers on.

The augmented reality branch is in progress and will need an interface
to bridge to normal camera when unavailable